Move YAML loading functions to new yaml_provider.py
