+++ activity/activity_PMCDeposit.py
@@ -9,7 +9,13 @@ import provider.s3lib as s3lib
 from provider.article_structure import ArticleInfo
 from provider.execution_context import get_session
 from provider.storage_provider import storage_context
-from provider import article_processing, downstream, email_provider, lax_provider, utils
+from provider import (
+    article_processing,
+    email_provider,
+    lax_provider,
+    utils,
+    yaml_provider,
+)
 from provider.ftp import FTP
 from activity.objects import Activity
 
@@ -109,7 +115,7 @@ class activity_PMCDeposit(Activity):
             self.directories.get("ZIP_DIR"), self.zip_file_name
         )
 
-        rules = downstream.load_config(self.settings)
+        rules = yaml_provider.load_config(self.settings)
         workflow_rules = rules.get("PMC")
 
         # repackage the archive zip into PMC zip format
